# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Gold Binar Doc",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": ""
      }
    ],
...
    "paths": {
...
      "/items/add-item": {
        "post": {
          "tags": [
            "Item"
          ],
          "summary": "Add Item",
          "requestBody": {
            "content": {
              "application/x-www-form-urlencoded": {
                "schema": {
                  "properties": {
                    "item_name": {
                      "type": "string",
                      "example": "sepatu"
                    },
                    "stock": {
                      "type": "integer",
                      "example": "3"
                    },
                    "price": {
                      "type": "integer",
                      "example": "75000"
                    }
                  }
                }
              }
            }
          },
          "parameters": [
            {
              "name": "token",
              "in": "header",
              "schema": {
                "type": "string"
              },
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyaWQiOjcsImlhdCI6MTY3OTExNDM0Nn0.hzGiGbgKpmFAc6zjBGzUBboPxyOrRgg_HaGjAYYKsVM"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {}
              }
            }
          }
        }
      },
      "/items/update-item": {
        "put": {
          "tags": [
            "Item"
          ],
          "summary": "Update Item",
          "requestBody": {
            "content": {
              "application/x-www-form-urlencoded": {
                "schema": {
                  "properties": {
                    "item_id": {
                      "type": "integer",
                      "example": "1"
                    },
                    "item_name": {
                      "type": "string",
                      "example": "sabun mandi wangi"
                    },
                    "stock": {
                      "type": "integer",
                      "example": "17"
                    },
                    "price": {
                      "type": "integer",
                      "example": "1750"
                    },
                    "status": {
                      "type": "integer",
                      "description": "1 = Active , 0 = Inactive",
                      "example": "1"
                    }
                  }
                }
              }
            }
          },
          "parameters": [
            {
              "name": "token",
              "in": "header",
              "schema": {
                "type": "string"
              },
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyaWQiOjcsImlhdCI6MTY3OTExNDM0Nn0.hzGiGbgKpmFAc6zjBGzUBboPxyOrRgg_HaGjAYYKsVM"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {}
              }
            }
          }
        }
      },
...
    }
  }